Move components to pods
[ci skip]
[#LEI-315]
