+++ exp-player/addon/components/exp-audioplayer/template.hbs
