+++ exp-player/addon/components/exp-blank/component.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from '../templates/components/exp-blank';
+import layout from './template';
 
 export default Ember.Component.extend({
     layout
