+++ exp-player/addon/components/exp-consent/template.hbs
