+++ exp-player/addon/components/exp-frame-base/component.js
