+++ exp-player/addon/components/exp-lookit-overview/template.hbs
