+++ exp-player/addon/components/exp-physics-pre-video/component.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 
 import ExpFrameBaseComponent from 'exp-player/components/exp-frame-base';
-import layout from 'exp-player/templates/components/exp-physics-pre-video';
+import layout from './template';
 
 export default ExpFrameBaseComponent.extend({
     layout: layout,
