+++ exp-player/addon/components/exp-player/template.hbs
