+++ exp-player/addon/components/exp-video-config/template.hbs
