+++ exp-player/addon/components/exp-video-consent/template.hbs
