+++ exp-player/addon/components/exp-video-preview/template.hbs
