+++ exp-player/addon/components/exp-video-record/component.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from '../templates/components/exp-video-record';
+import layout from './template';
 
 import ExpFrameBaseComponent from 'exp-player/components/exp-frame-base';
 // import FullScreen from '../mixins/full-screen';
