+++ exp-player/addon/components/exp-video-record/template.hbs
