+++ exp-player/addon/components/exp-video-setup/component.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 
 import ExpFrameBaseComponent from 'exp-player/components/exp-frame-base';
-import layout from '../templates/components/exp-video-setup';
+import layout from './template';
 
 export default ExpFrameBaseComponent.extend({
     type: 'exp-video-setup',
