+++ exp-player/addon/components/exp-video-setup/template.hbs
