+++ exp-player/addon/components/exp-video/component.js
@@ -1,4 +1,4 @@
-import layout from '../templates/components/exp-video';
+import layout from './template';
 
 import ExpFrameBaseComponent from 'exp-player/components/exp-frame-base';
 import FullScreen from '../mixins/full-screen';
